+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## 5.0.x-dev
 ### Added
 - Support for PHP 7.0.x
+- Source files were moved form `lib` to `src`
 
 ### Removed
 - Support for PHP 5.x